requirements.txt
# django
Django==4.0
django-extensions==3.1.5
django-ordered-model==3.4.3
django-redis-cache==3.0.0

# djangorestframework
djangorestframework==3.13.1
django-filter==21.1
django-dry-rest-permissions==1.2.0  # https://github.com/FJNR-inc/dry-rest-permissions

# psycopg2
psycopg2-binary==2.9.2


# tests, coverage
model-bakery==1.3.3
coverage==6.2

# celery
celery==5.2.1
django-celery-results==2.2.0
django-celery-beat==2.2.1


# dependency and safety checker
safety==1.10.3
insecure-package==0.1.0
graphviz==0.19.1
pipdeptree==2.2.0

# pre-commit, linter, formatter
pre_commit==2.16.0
flake8==4.0.1
flake8-bugbear==21.11.29
isort==5.10.1
autoflake==1.4
black==21.12b0